Counting chat turns now work, had to refactor for container/ports to get it working
