+++ docker-compose.yml
@@ -15,6 +15,8 @@ services:
     command: python -m spanishtutor.src.main
     ports:
       - "7860:7860"
+    environment:
+      - API_URL=http://spanish-tutor-api:8000/chat # for running with docker compose (for docker internal dns name)
     depends_on:
       - spanish-tutor-api
